requirements.txt
streamlit
python-dotenv
openai
langchain
PyPDF2
chromadb
faiss-cpu
pdf2image
youtube_transcript_api